@@ -1,12 +1,11 @@
-﻿namespace Saket.Serialization
+﻿namespace Saket.Serialization;
+
+/// <summary>
+/// Interface for implementing custom serializable types.
+/// Both reading and writing is combined in a single function.
+/// You can know by reading ISerializer.IsReader
+/// </summary>
+public interface ISerializable
 {
-    /// <summary>
-    /// Interface for implementing custom serializable types.
-    /// Both reading and writing is combined in a single function.
-    /// You can know by reading ISerializer.IsReader
-    /// </summary>
-    public interface ISerializable
-    {
-        void Serialize(ISerializer serializer);
-    }
+    void Serialize(ISerializer serializer);
 }
