@@ -1,122 +1,196 @@
 ﻿using System;
-using System.Numerics;
+using System.Collections.Generic;
 
-namespace Saket.Serialization
+namespace Saket.Serialization;
+
+/// <summary>
+/// LoadBytes must be called before any serialize functions.
+/// </summary>
+public interface ISerializer
 {
+    public long Position { get; set; }
     /// <summary>
-    /// LoadBytes must be called before any serialize functions.
+    /// Whether is serializer is a reader.
     /// </summary>
-    public interface ISerializer
+    public bool IsReader { get; }
+    /// <summary>
+    /// Whether is serializer is a writer. 
+    /// </summary>
+    public bool IsWriter => !IsReader;
+
+    /// <summary>
+    /// Ensure that the number of bytes is serializable. 
+    /// This method must be called before any serialize functions to ensure bounds.
+    /// </summary>
+    /// <param name="count"></param>
+    public bool LoadBytes(int count);
+
+
+    //---- Bytes ----
+    public void Serialize(ref Span<byte> value, ForBytes unused = default);
+    public unsafe void Serialize(byte* value, int count)
     {
-        public long Position { get; set; }
-        /// <summary>
-        /// Whether is serializer is a reader.
-        /// </summary>
-        public bool IsReader { get; }
-        /// <summary>
-        /// Whether is serializer is a writer. 
-        /// </summary>
-        public bool IsWriter => !IsReader;
+        var span = new Span<byte>(value, count);
+        Serialize(ref span);
+    }
 
-        /// <summary>
-        /// Ensure that the number of bytes is serializable. 
-        /// This method must be called before any serialize functions to ensure bounds.
-        /// </summary>
-        /// <param name="count"></param>
-        public bool LoadBytes(int count);
+    // ---- String ----
+    public void Serialize(ref string value)
+    {
+        int length = value.Length;
+        Serialize(ref length);
 
-        /// <summary>
-        /// Serialize arbitrary number of bytes 
-        /// </summary>
-        /// <param name="length"></param>
-        /// <returns></returns>
-        public void SerializeBytes(ref byte[] value, int offset = 0, int count = 0);
-        /// <summary>
-        /// Serialize arbitrary number of bytes from ptr
-        /// </summary>
-        /// <param name="length"></param>
-        /// <returns></returns>
-        public unsafe void SerializeBytes(byte* value, int count);
+        Span<char> chars = stackalloc char[length];
+        if(IsWriter)
+            value.AsSpan().CopyTo(chars);
 
-        /*
-        public void SerializeBoolean(ref bool value);
-
-        public void SerializeByte(ref byte value);
-        public void SerializeSByte(ref sbyte value);
-
-        public void SerializeChar(ref char value);
-
-        /// <summary>
-        /// The maximum length of the string is dependent on the implementation
-        /// </summary>
-        /// <param name="value"></param>
-        public void SerializeString(ref string value);
-
-        public void SerializeUInt16(ref ushort value);
-        public void SerializeUInt32(ref uint value);
-        public void SerializeUInt64(ref ulong value);
-
-        public void SerializeInt16(ref short value);
-        public void SerializeInt32(ref int value);
-        public void SerializeInt64(ref long value);
-
-        public void SerializeSingle(ref float value);
-        public void SerializeDouble(ref double value);
-        public void SerializeDecimal(ref decimal value);
-
-        public void SerializeVector2(ref Vector2 value);
-        public void SerializeVector3(ref Vector3 value);
-        public void SerializeVector4(ref Vector4 value);
-        public void SerializeMatrix4x4(ref Matrix4x4 value);
-        public void SerializeQuaternion(ref Quaternion value);*/
-
-        //---- Generic ----
-        public void Serialize<T>(ref T value) where T : unmanaged;
-        public void Serialize<T>(ref T[] value) where T : unmanaged
+        for (int i = 0; i < length; i++)
         {
-            unsafe
-            {
-                int length = 0;
-                Serialize(ref length);
-                if (value.Length < length)
-                    value = new T[length];
+            Serialize(ref chars[i]);
+        }
 
-                for (int i = 0; i < length; i++)
-                {
-                    Serialize<T>(ref value[i]);
-                }
+        value = new string(chars);
+    }
+
+    // ---- Generic Unmanged ----
+    public void Serialize<T>(ref T value, ForUnmanaged unused = default) where T : unmanaged;
+    public void Serialize<T>(ref T[] value, ForUnmanaged unused = default) where T : unmanaged
+    {
+        int length = value.Length;
+        Serialize(ref length);
+        if (value.Length < length)
+            value = new T[length];
+
+        for (int i = 0; i < length; i++)
+        {
+            Serialize(ref value[i]);
+        }
+    }
+    public void Serialize<T>(ref List<T> list, ForUnmanaged unused = default) where T : unmanaged
+    {
+        int length = list.Count;
+        Serialize(ref length);
+        
+        list.Capacity = length;
+
+        // Length now contains the length of the dictionary
+        for (int i = 0; i < length; i++)
+        {
+            T value = default;
+            Serialize(ref value);
+            list[i] = value;
+        }
+    }
+    public void Serialize<K, V>(ref Dictionary<K, V> dictionary, ForUnmanaged unused = default)
+         where K : unmanaged
+         where V : unmanaged
+    {
+        int length = dictionary.Count;
+        Serialize(ref length);
+        // Length now contains the length of the dictionary
+        if (IsReader)
+        {
+            // Clear the old one
+            dictionary.Clear();
+            for (int i = 0; i < length; i++)
+            {
+                K key = default;
+                Serialize(ref key);
+
+                V value = default;
+                Serialize(ref value);
+
+                 dictionary.Add(key, value);
             }
         }
-
-        public void SerializeEnum<T>(ref T value) where T : unmanaged, Enum;
-
-        public void SerializeSerializable<T>(ref T value) where T : ISerializable, new()
-        { 
-            value.Serialize(this); 
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="value"></param>
-        public void SerializeSerializable<T>(ref T[] value) where T : ISerializable, new()
+        else
         {
-            unsafe
+            foreach (var item in dictionary)
             {
-                int length = 0;
-                Serialize(ref length);
+                K key = item.Key;
+                Serialize(ref key);
 
-                if (value.Length < length)
-                {
-                    value = new T[length];
-                }
-
-                for (int i = 0; i < length; i++)
-                {
-                    SerializeSerializable<T>(ref value[i]);
-                }
+                V value = item.Value;
+                Serialize(ref value);
             }
         }
     }
+
+    // ---- Enum ----
+    public void Serialize<T>(ref Enum value, ForEnum unused = default) where T : Enum;
+
+    // ---- Serializable ----
+    public void Serialize<T>(ref T value) where T : ISerializable, new()
+    { 
+        value.Serialize(this); 
+    }
+    public void Serialize<T>(ref T[] value) where T : ISerializable, new()
+    {
+        int length = value.Length;
+        Serialize(ref length);
+
+        if (value.Length < length)
+        {
+            value = new T[length];
+        }
+
+        for (int i = 0; i < length; i++)
+        {
+            Serialize(ref value[i]);
+        }
+    }
+    public void Serialize<T>(ref List<T> list) where T : ISerializable, new()
+    {
+        int length = list.Count;
+        Serialize(ref length);
+
+        list.Capacity = length;
+
+        // Length now contains the length of the dictionary
+        for (int i = 0; i < length; i++)
+        {
+            T value = new();
+            Serialize(ref value);
+            list[i] = value;
+        }
+    }
+    public void Serialize<K, V>(ref Dictionary<K, V> dictionary)
+        where K : unmanaged
+        where V : ISerializable, new()
+    {
+        int length = dictionary.Count;
+        Serialize(ref length);
+        // Length now contains the length of the dictionary
+        if (IsReader)
+        {
+            // Clear the old one
+            dictionary.Clear();
+            for (int i = 0; i < length; i++)
+            {
+                K key = default;
+                Serialize(ref key);
+
+                V value = new();
+                Serialize(ref value);
+
+                dictionary.Add(key, value);
+            }
+        }
+        else
+        {
+            foreach (var item in dictionary)
+            {
+                K key = item.Key;
+                Serialize(ref key);
+
+                V value = item.Value;
+                Serialize(ref value);
+            }
+        }
+    }
+
+    struct ForBytes { }
+    struct ForUnmanaged { }
+    struct ForEnum { }
+    struct ForSerializable { }
 }
